requirement.txt
numpy==1.19.5
pandas==1.2.2
scikit-learn==0.24.1
matplotlib==3.3.4
jupyter==1.0.0
opencv-python==4.5.1.48
tensorflow-cpu==2.4.0
pillow==8.4.0
tensorflow==2.7.0
keras==2.7.0
scikit-learn==1.0.1
gradio==2.3.6